+++ gamevault/ViewModels/SettingsViewModel.cs
@@ -130,7 +130,7 @@ namespace gamevault.ViewModels
         {
             get
             {
-                return "1.2.0";
+                return "1.2.1";
             }
         }